Version 1.0
I think we could compete with this code.
Implemented
- Cheesy Drive (Ask Allen how Cheesy drive works, the driver triggers
are now turning not collection.)
- Drive Mode Selector (Halo and Cheesy are the only two options right
now)
- Changed Joystick deadzones and sensitivity should be much smoother
now, a lot more control
- Can adjust Cheesy sensitivity from dashboard we'll need to tune it.

Tilt motor locks when you push the left stick button. Can adjust the
lock value but so far .3 is about right

Shooter speeds are set from the dashboard when you press that A button
so we can optiomize shoots

Gamepad class button assignments are fixed

Lots more dashboard items
- Disabled toggle shows that SmartDashboard is updating even when
disabled, it wasn't for a bit
- add a Utlitities class so we can add methods easily for the whole
program